move more functions to utils
